requirements.txt
opencv-python
numpy
pychalk
Pillow
git+https://github.com/IsmaeelAkram/pycamera.git